Developer

+++ src/main/java/com/travelapp/services/TicketService.java
@@ -1,8 +1,11 @@
 package com.travelapp.services;
 
+import com.travelapp.exceptions.BadRequestException;
 import com.travelapp.exceptions.ResourceNotFoundException;
+import com.travelapp.exceptions.ResourcePersistenceException;
 import com.travelapp.models.Ticket;
 import com.travelapp.repos.TicketRepository;
+import com.travelapp.util.Validator;
 import com.travelapp.web.dtos.TicketDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,6 +13,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import static com.travelapp.util.Validator.*;
 
 @Service
 public class TicketService {
@@ -39,27 +44,59 @@ public class TicketService {
 
     @Transactional(readOnly=true)
     public TicketDto getById(int id) {
-        return new TicketDto(ticketRepo.findById(id));
+        if(!isValidId(id)){
+            throw new BadRequestException();
+        }
+        try {
+            return new TicketDto(ticketRepo.findById(id));
+        }
+        catch (Exception e) {
+            throw new ResourceNotFoundException();
+        }
+
     }
 
     @Transactional
     public TicketDto save(TicketDto ticketDto) {
-
-        System.out.println(ticketDto.getArrivalTime());
-        System.out.println(ticketDto.getDepartureTime());
+        if(!isValidTicket(ticketDto)){
+            throw new BadRequestException();
+        }
         Ticket ticket = new Ticket(ticketDto);
-        return new TicketDto(ticketRepo.save(ticket));
+        try {
+            return new TicketDto(ticketRepo.save(ticket));
+        }
+        catch (Exception e) {
+            throw new ResourcePersistenceException();
+        }
+
 
     }
 
     @Transactional
     public boolean update(Ticket ticket) {
-        return ticketRepo.update(ticket);
+        if(!isValidUpdateTicket(ticket)){
+            throw new BadRequestException();
+        }
+        try {
+            return ticketRepo.update(ticket);
+        }
+        catch (Exception e) {
+            throw new ResourcePersistenceException("Ticket could not be updated");
+        }
+
     }
 
     @Transactional
     public boolean deleteById(int id) {
-        return ticketRepo.deleteById(id);
+        if(!isValidId(id)){
+            throw new BadRequestException();
+        }
+        try {
+            return ticketRepo.deleteById(id);
+        }
+        catch (Exception e) {
+            return true;
+        }
     }
 
 }
